+++ src/main/kotlin/org/simplemc/simpleannounce/config/SimpleAnnounceConfig.kt
@@ -4,8 +4,10 @@ import com.fasterxml.jackson.annotation.JsonAlias
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.fasterxml.jackson.annotation.JsonTypeInfo
 import com.fasterxml.jackson.annotation.JsonUnwrapped
+import org.bukkit.Sound
 import org.bukkit.boss.BarColor
 import org.bukkit.boss.BarStyle
+import org.simplemc.simpleannounce.config.SimpleAnnounceConfig.AnnouncementConfig.Chat.ChatMessage
 import org.simplemc.simpleannounce.inTicks
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
@@ -16,8 +18,8 @@ data class SimpleAnnounceConfig(
     val announcements: List<AnnouncementConfig<*>>,
 ) {
     companion object {
-        private fun Duration?.checkNullZeroOrPositive(name: String) {
-            check(this == null || this == Duration.Companion.ZERO || this.isPositive()) {
+        private fun Duration?.requireNullZeroOrPositive(name: String) {
+            require(this == null || this == Duration.Companion.ZERO || this.isPositive()) {
                 "When set, $name must be >= 0s"
             }
         }
@@ -27,19 +29,24 @@ data class SimpleAnnounceConfig(
     val autoReloadTicks = autoReload?.inTicks
 
     init {
-        check(autoReload == null || autoReload == Duration.Companion.ZERO || autoReload.inWholeMinutes >= 1) {
+        require(autoReload == null || autoReload == Duration.Companion.ZERO || autoReload.inWholeMinutes >= 1) {
             "When set, Auto Reload Duration must be > 1 minute"
         }
     }
 
     @JsonTypeInfo(use = JsonTypeInfo.Id.SIMPLE_NAME, include = JsonTypeInfo.As.PROPERTY, property = "type")
-    sealed class AnnouncementConfig<T> {
+    sealed class AnnouncementConfig<T : AnnouncementConfig.Message> {
         abstract val random: Boolean
         abstract val delay: Duration
         abstract val repeat: Duration?
+        abstract val sound: SoundConfig?
         abstract val includesPermissions: List<String>
         abstract val excludesPermissions: List<String>
         abstract val messages: List<T>
+
+        interface Message {
+            val sound: SoundConfig?
+        }
 
         @JsonIgnore
         val delayTicks = delay.inTicks.toInt()
@@ -48,23 +55,27 @@ data class SimpleAnnounceConfig(
         val repeatTicks = repeat?.inTicks?.toInt()
 
         init {
-            delay.checkNullZeroOrPositive("delay")
-            repeat.checkNullZeroOrPositive("repeat")
+            delay.requireNullZeroOrPositive("delay")
+            repeat.requireNullZeroOrPositive("repeat")
         }
 
         data class Chat(
             override val random: Boolean = false,
             override val delay: Duration = Duration.Companion.ZERO,
             override val repeat: Duration? = null,
+            override val sound: SoundConfig? = null,
             override val includesPermissions: List<String> = emptyList(),
             override val excludesPermissions: List<String> = emptyList(),
-            @field:JsonAlias("message") override val messages: List<String>,
-        ) : AnnouncementConfig<String>()
+            @field:JsonAlias("message") override val messages: List<ChatMessage>,
+        ) : AnnouncementConfig<ChatMessage>() {
+            data class ChatMessage(val message: String, override val sound: SoundConfig? = null) : Message
+        }
 
         data class Boss(
             override val random: Boolean = false,
             override val delay: Duration = Duration.Companion.ZERO,
             override val repeat: Duration? = null,
+            override val sound: SoundConfig? = null,
             override val includesPermissions: List<String> = emptyList(),
             override val excludesPermissions: List<String> = emptyList(),
             @field:JsonAlias("message") override val messages: List<BossBarMessage>,
@@ -72,8 +83,9 @@ data class SimpleAnnounceConfig(
         ) : AnnouncementConfig<Boss.BossBarMessage>() {
             data class BossBarMessage(
                 val message: String,
+                override val sound: SoundConfig? = null,
                 @field:JsonUnwrapped val barConfig: BarConfig? = null,
-            ) {
+            ) : Message {
                 init {
                     require(message.length <= 64) { "Boss Bar text must be <= 64 characters" }
                 }
@@ -90,7 +102,7 @@ data class SimpleAnnounceConfig(
                 val holdTicks = hold.inTicks
 
                 init {
-                    hold.checkNullZeroOrPositive("hold")
+                    hold.requireNullZeroOrPositive("hold")
                 }
             }
         }
@@ -99,6 +111,7 @@ data class SimpleAnnounceConfig(
             override val random: Boolean = false,
             override val delay: Duration = Duration.Companion.ZERO,
             override val repeat: Duration? = null,
+            override val sound: SoundConfig? = null,
             override val includesPermissions: List<String> = emptyList(),
             override val excludesPermissions: List<String> = emptyList(),
             @field:JsonAlias("message") override val messages: List<TitleMessage>,
@@ -107,8 +120,9 @@ data class SimpleAnnounceConfig(
             data class TitleMessage(
                 val title: String,
                 val subtitle: String? = null,
+                override val sound: SoundConfig? = null,
                 @field:JsonUnwrapped val titleConfig: TitleConfig? = null,
-            )
+            ) : Message
 
             data class TitleConfig(
                 val fadeIn: Duration = 500.milliseconds,
@@ -125,10 +139,17 @@ data class SimpleAnnounceConfig(
                 val fadeOutTicks = fadeOut.inTicks.toInt()
 
                 init {
-                    fadeIn.checkNullZeroOrPositive("fadeIn")
-                    stay.checkNullZeroOrPositive("stay")
-                    fadeOut.checkNullZeroOrPositive("fadeOut")
+                    fadeIn.requireNullZeroOrPositive("fadeIn")
+                    stay.requireNullZeroOrPositive("stay")
+                    fadeOut.requireNullZeroOrPositive("fadeOut")
                 }
+            }
+        }
+
+        data class SoundConfig(val sound: Sound, val volume: Float = 1F, val pitch: Float = 1F) {
+            init {
+                require(volume >= 0 && volume <= 1) { "Sound volume must be between 0 and 1" }
+                require(pitch >= 0.5 && pitch <= 2) { "Sound pitch must be between 0.5 and 2" }
             }
         }
     }