+++ src/main/kotlin/org/simplemc/simpleannounce/sender/TitleSender.kt
@@ -1,6 +1,5 @@
 package org.simplemc.simpleannounce.sender
 
-import org.bukkit.Bukkit
 import org.bukkit.plugin.Plugin
 import org.simplemc.simpleannounce.config.SimpleAnnounceConfig.AnnouncementConfig.Title
 
@@ -9,20 +8,17 @@ class TitleSender(
     announcement: Title,
 ) : AnnouncementSender<Title.TitleMessage, Title>(plugin, announcement) {
     override fun run() {
-        val message = getNextAnnouncement()
+        val message = getNextMessage()
         val titleConfig = message.titleConfig ?: announcement.titleConfig
 
-        Bukkit.getOnlinePlayers()
-            .filterNotNull()
-            .filter(this::shouldSendTo)
-            .forEach {
-                it.sendTitle(
-                    message.title,
-                    message.subtitle,
-                    titleConfig.fadeInTicks,
-                    titleConfig.stayTicks,
-                    titleConfig.fadeOutTicks,
-                )
-            }
+        send(message) {
+            it.sendTitle(
+                message.title,
+                message.subtitle,
+                titleConfig.fadeInTicks,
+                titleConfig.stayTicks,
+                titleConfig.fadeOutTicks,
+            )
+        }
     }
 }